extract quick open paging
