+++ src/vs/workbench/parts/extensions/electron-browser/extensionsQuickOpen.ts
@@ -13,6 +13,7 @@ import * as dom from 'vs/base/browser/dom';
 import Severity from 'vs/base/common/severity';
 import { onUnexpectedError } from 'vs/base/common/errors';
 import { IAutoFocus, Mode, IModel, IDataSource, IRenderer, IRunner, IContext, IAccessiblityProvider } from 'vs/base/parts/quickopen/common/quickOpen';
+import { QuickOpenPagedModel, IPagedRenderer } from 'vs/base/parts/quickopen/common/quickOpenPaging';
 import { matchesContiguousSubString } from 'vs/base/common/filters';
 import { QuickOpenHandler } from 'vs/workbench/browser/quickopen';
 import { IHighlight } from 'vs/base/parts/quickopen/browser/quickOpenModel';
@@ -178,10 +179,6 @@ class AccessibilityProvider implements IAccessiblityProvider<IExtensionEntry> {
 	public getAriaLabel(entry: IExtensionEntry): string {
 		return nls.localize('extensionAriaLabel', "{0}, {1}, extensions picker", entry.extension.displayName, entry.extension.description);
 	}
-}
-
-interface IPagedRenderer<T> extends IRenderer<T> {
-	renderPlaceholder(index: number, templateId: string, data: any): void;
 }
 
 class Renderer implements IPagedRenderer<IExtensionEntry> {
@@ -403,124 +400,6 @@ export class LocalExtensionsHandler extends QuickOpenHandler {
 	}
 }
 
-interface IStubTemplateData<T> {
-	data: T;
-	disposable: IDisposable;
-}
-
-interface IStub {
-	index: number;
-}
-
-class PagedRenderer<T> implements IRenderer<IStub> {
-
-	constructor(
-		private model: PagedModel<T>,
-		private renderer: IPagedRenderer<T>
-	) {}
-
-	getHeight(stub: IStub): number {
-		return 48; // TODO
-	}
-
-	getTemplateId(stub: IStub): string {
-		return 'extension'; // TODO
-	}
-
-	renderTemplate(templateId: string, container: HTMLElement): IStubTemplateData<T> {
-		const data = this.renderer.renderTemplate(templateId, container);
-		return { data, disposable: { dispose: () => {} } };
-	}
-
-	renderElement({ index }: IStub, templateId: string, data: IStubTemplateData<T>): void {
-		data.disposable.dispose();
-
-		if (this.model.isResolved(index)) {
-			return this.renderer.renderElement(this.model.get(index), templateId, data.data);
-		}
-
-		let didUnrender = false;
-		data.disposable = { dispose: () => didUnrender = true };
-		this.renderer.renderPlaceholder(index, templateId, data.data);
-
-		this.model.resolve(index).done(entry => {
-			if (didUnrender) {
-				return;
-			}
-
-			this.renderer.renderElement(entry, templateId, data.data);
-		});
-	}
-
-	disposeTemplate(templateId: string, data: IStubTemplateData<T>): void {
-		data.disposable.dispose();
-		data.disposable = null;
-		this.renderer.disposeTemplate(templateId, data.data);
-	}
-}
-
-class PagedDataSource<T> implements IDataSource<IStub> {
-
-	constructor(
-		private model: PagedModel<T>,
-		private dataSource: IDataSource<T>
-	) {}
-
-	getId({ index }: IStub): string {
-		return `paged-${ index }`;
-	}
-
-	getLabel({ index }: IStub): string {
-		return this.model.isResolved(index) ? this.dataSource.getLabel(this.model.get(index)) : '';
-	}
-}
-
-class PagedRunner<T> implements IRunner<IStub> {
-
-	constructor(
-		private model: PagedModel<T>,
-		private runner: IRunner<T>
-	) {}
-
-	run({ index }: IStub, mode: Mode, context: IContext): boolean {
-		if (this.model.isResolved(index)) {
-			return this.runner.run(this.model.get(index), mode, context);
-		}
-
-		return false;
-	}
-}
-
-class PagedModel2<T> implements IModel<IStub> {
-
-	public dataSource: IDataSource<IStub>;
-	public renderer: IRenderer<IStub>;
-	public runner: IRunner<IStub>;
-	// public filter: IFilter<IStub>;
-	// public accessibilityProvider: IAccessiblityProvider<IStub>;
-	public entries: IStub[];
-
-	constructor(
-		model: PagedModel<T>,
-		dataSource: IDataSource<T>,
-		renderer: IPagedRenderer<T>,
-		runner: IRunner<T>
-		// filter?: IFilter<T>,
-		// accessibilityProvider?: IAccessiblityProvider<T>
-	) {
-		this.dataSource = new PagedDataSource(model, dataSource);
-		this.renderer = new PagedRenderer(model, renderer);
-		this.runner = new PagedRunner(model, runner);
-		// this.filter = new PagedFilter(model, filter);
-		// this.accessibilityProvider = new PagedAccessibilityProvider(model, accessibilityProvider);
-
-		this.entries = [];
-		for (let index = 0, len = model.length; index < len; index++) {
-			this.entries.push({ index });
-		}
-	}
-}
-
 class GalleryExtensionsModel implements IModel<IExtensionEntry> {
 
 	public dataSource = new DataSource();
@@ -573,21 +452,20 @@ export class GalleryExtensionsHandler extends QuickOpenHandler {
 	}
 
 	getResults(text: string): TPromise<IModel<number>> {
-		return this.delayer.trigger(() => this.galleryService.query({ text }))
-			.then((result: IQueryResult) => {
-				const pager = mapPager(result, extension => ({
-					extension,
-					highlights: getHighlights(text.trim(), extension, false),
-					state: ExtensionState.Uninstalled
-				}));
+		return this.delayer.trigger(() => this.galleryService.query({ text })).then((result: IQueryResult) => {
+			const pager = mapPager(result, extension => ({
+				extension,
+				highlights: getHighlights(text.trim(), extension, false),
+				state: ExtensionState.Uninstalled
+			}));
 
-				return new PagedModel2<IExtensionEntry>(
-					new PagedModel(pager),
-					new DataSource(),
-					this.instantiationService.createInstance(Renderer),
-					this.instantiationService.createInstance(InstallRunner)
-				);
-			});
+			return new QuickOpenPagedModel<IExtensionEntry>(
+				new PagedModel(pager),
+				new DataSource(),
+				this.instantiationService.createInstance(Renderer),
+				this.instantiationService.createInstance(InstallRunner)
+			);
+		});
 	}
 
 	getEmptyLabel(input: string): string {
